requirements.txt
django
django-rest-framework
djangorestframework_simplejwt